fix url booking manager
