--- src/java/bookcontroller/ManageBookingServlet.java
@@ -1,33 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package bookcontroller;
-
-import DAO.BookDAO;
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.List;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import model.Book;
-
-/**
- *
- * @author Hoang Phuc
- */
-public class ManageBookingServlet extends HttpServlet {
-
-    protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        PrintWriter pr = response.getWriter();
-        BookDAO u = new BookDAO();
-        List<Book> lst = u.getBookItem();
-        request.setAttribute("bookinglist", lst);
-        request.getRequestDispatcher("ManageBooking.jsp").forward(request, response);
-    }
-}
